requirements.txt
flask==3.1.0
flask-cors==5.0.1
pydantic==2.11.3
python-dotenv==1.1.0
Werkzeug==3.1.3
Flask-PyMongo==3.0.1
flask-login==0.6.3
bcrypt==4.1.2
